+++ example/lib/example.dart
@@ -34,42 +34,43 @@ void main(List<String> arguments) async {
   print('Requested Video: ${video.toString()}');
 
   // Get a bunch of videos and print them
-  final PaginatedResult<VideoFull> videoList = await holodexClient.getVideos(
-      // channelId: 'UCsYcCwDqv6Sg8KMIIMF54SA', // Kiriku Translation
-      includes: <Includes>[
-        Includes.channelStats,
-        Includes.clips,
-        // Includes.description,
-        Includes.liveInfo,
-        Includes.mentions,
-        Includes.refers,
-        Includes.simulcasts,
-        Includes.songs,
-        Includes.sources,
-      ],
-      languages: <Language>[Language.all],
-      limit: 50,
-      maxUpcomingHours: 1000,
-      // mentionedChannelId: 'UCDqI2jOz0weumE8s7paEk6g', // Roboco
-      offset: 50,
-      order: Order.descending,
-      // organization: Organization.Hololive,
-      paginated: true,
-      videoSort: <VideoSort>[VideoSort.availableAt],
-      videoStatus: <VideoStatus>[VideoStatus.past],
-      // Videos of type VideoType.clip cannot not have topic. Streams may or may not have topic.
-      // topicId: 'singing',
-      videoType: VideoType.all);
-  print(
-      'Videos: ${videoList.items.length}\nTotal Videos: ${videoList.total}\n');
+  final videoFilter = VideoFilter(
+    // channelId: 'UCsYcCwDqv6Sg8KMIIMF54SA', // Kiriku Translation
+    includes: <Includes>[
+      Includes.channelStats,
+      Includes.clips,
+      // Includes.description,
+      Includes.liveInfo,
+      Includes.mentions,
+      Includes.refers,
+      Includes.simulcasts,
+      Includes.songs,
+      Includes.sources,
+    ],
+    languages: <Language>[Language.all],
+    limit: 50,
+    maxUpcomingHours: 1000,
+    // mentionedChannelId: 'UCDqI2jOz0weumE8s7paEk6g', // Roboco
+    offset: 50,
+    order: Order.descending,
+    // organization: Organization.Hololive,
+    paginated: true,
+    videoSort: <VideoSort>[VideoSort.availableAt],
+    videoStatus: <VideoStatus>[VideoStatus.past],
+    // Videos of type VideoType.clip cannot not have topic. Streams may or may not have topic.
+    // topicId: 'singing',
+    videoType: VideoType.all,
+  );
+  
+  final PaginatedResult<VideoFull> videoList = await holodexClient.getVideos(videoFilter);
+  print('Videos: ${videoList.items.length}\nTotal Videos: ${videoList.total}\n');
 
   // // Get live videos
   final PaginatedResult<VideoFull> liveVideos =
       await holodexClient.getLiveVideos(includes: [Includes.channelStats]);
   print('Live videos: ${liveVideos.items.length}\n');
 
-  final ceresFauna =
-      await holodexClient.getChannelFromId('UCO_aKKYxn4tvrqPjcTzZ6EQ');
+  final ceresFauna = await holodexClient.getChannelFromId('UCO_aKKYxn4tvrqPjcTzZ6EQ');
   print('Requested Channel Name: ${ceresFauna.name}\n');
 
   final List<Channel> channels = await holodexClient.getChannels(
@@ -80,22 +81,20 @@ void main(List<String> arguments) async {
       channelSort: [ChannelSort.organization]);
   print('Atelier Live Channels: ${channels.length}\n');
 
-  final List<Video> quickLiveVideos =
-      await holodexClient.getLiveVideosFromChannelsQuickly([
+  final List<Video> quickLiveVideos = await holodexClient.getLiveVideosFromChannelsQuickly([
     'UCQ0UDLQCjY0rmuxCDE38FGg', // Matsuri
     'UCZlDXzGoo7d44bwdNObFacg', // Kanata
     'UCqm3BQLlJfvkTsX_hvm0UmA' // Watame
   ]);
   print('Requested Live Videos From Channels: ${quickLiveVideos.length}\n');
 
-  final PaginatedResult<VideoFull> matsuriClips = await holodexClient
-      .getVideosRelatedToChannel('UCQ0UDLQCjY0rmuxCDE38FGg', // Matsuri
+  final PaginatedResult<VideoFull> matsuriClips =
+      await holodexClient.getVideosRelatedToChannel('UCQ0UDLQCjY0rmuxCDE38FGg', // Matsuri
           type: VideoSearchType.clips);
   print('Clips including Matsuri: ${matsuriClips.total}');
   print('Returned clips including Matsuri: ${matsuriClips.items.length}\n');
 
-  final PaginatedResult<VideoFull> matsuriClips2 =
-      await holodexClient.getVTuberClips(
+  final PaginatedResult<VideoFull> matsuriClips2 = await holodexClient.getVTuberClips(
     'UCQ0UDLQCjY0rmuxCDE38FGg', // Matsuri
   );
   print('Clips including Matsuri: ${matsuriClips2.total}');
@@ -124,8 +123,7 @@ void main(List<String> arguments) async {
   print('Video Comments With Timestamps: ${timestampComments?.length}');
   print('Video Recommendations: ${recommendations?.length}\n');
 
-  final PaginatedResult<VideoFull> searchVideos =
-      await holodexClient.searchVideos(
+  final PaginatedResult<VideoFull> searchVideos = await holodexClient.searchVideos(
     searchSort: SearchSort.newest,
     languages: [Language.all],
     searchTargets: [SearchTarget.clip, SearchTarget.stream],
@@ -142,8 +140,7 @@ void main(List<String> arguments) async {
   );
   print('Videos Found: ${searchVideos.items.length}\n');
 
-  final PaginatedResult<VideoWithComments> searchComments =
-      await holodexClient.searchComments(
+  final PaginatedResult<VideoWithComments> searchComments = await holodexClient.searchComments(
     searchSort: SearchSort.newest,
     languages: [Language.all],
     searchTargets: [SearchTarget.clip, SearchTarget.stream],