+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **Breaking**: Renamed `HolodexClient` constructor parameter `httpClient` to `client`
 - **Breaking**: All uses of `startScheduled`, `startActual`, `endActual`, and `liveViewers`
 - **Breaking**: Changed `HolodexClient::searchComments` return type to `PaginatedResult<VideoWithComments>`
+- **Breaking**: Changed `HolodexClient::getVideos` parameters to an optional `VideoFilter`
 - **Breaking**: Removed `VideoWithCommentsList`
 - **Breaking**: Renamed `client.dart` to `holodex_client.dart`. Technically a breaking change, but only if
 implementation was imported.